bug fixes after refactoring
